refactor: extract `kms` module to `external_services` crate
