+++ crates/external_services/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "external_services"
+description = "Interactions of the router with external systems"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.65"
+readme = "README.md"
+license = "Apache-2.0"
+
+[features]
+kms = ["dep:aws-config", "dep:aws-sdk-kms"]
+
+[dependencies]
+aws-config = { version = "0.54.1", optional = true }
+aws-sdk-kms = { version = "0.24.0", optional = true }
+base64 = "0.21.0"
+error-stack = "0.3.1"
+once_cell = "1.17.1"
+serde = { version = "1.0.155", features = ["derive"] }
+thiserror = "1.0.39"
+tokio = "1.26.0"
+
+# First party crates
+common_utils = { version = "0.1.0", path = "../common_utils" }
+router_env = { version = "0.1.0", path = "../router_env", features = ["log_extra_implicit_fields", "log_custom_entries_to_extra"] }
