+++ crates/external_services/README.md
@@ -1,0 +1,4 @@
+# External Services
+
+This crate includes interactions with external systems, which may be shared
+across crates within this workspace.
