+++ crates/router/Cargo.toml
@@ -10,8 +10,8 @@ license = "Apache-2.0"
 build = "src/build.rs"
 
 [features]
-default = ["kv_store", "stripe", "oltp", "olap","accounts_cache"]
-kms = ["aws-config", "aws-sdk-kms"]
+default = ["kv_store", "stripe", "oltp", "olap", "accounts_cache"]
+kms = ["external_services/kms"]
 basilisk = ["kms"]
 stripe = ["dep:serde_qs"]
 sandbox = ["kms", "stripe", "basilisk"]
@@ -21,7 +21,7 @@ production = []
 kv_store = []
 accounts_cache = []
 openapi = ["olap", "oltp"]
-vergen = [ "router_env/vergen" ]
+vergen = ["router_env/vergen"]
 
 
 [dependencies]
@@ -31,8 +31,6 @@ actix-rt = "2.8.0"
 actix-web = "4.3.1"
 async-bb8-diesel = { git = "https://github.com/juspay/async-bb8-diesel", rev = "9a71d142726dbc33f41c1fd935ddaa79841c7be5" }
 async-trait = "0.1.66"
-aws-config = { version = "0.54.1", optional = true }
-aws-sdk-kms = { version = "0.24.0", optional = true }
 base64 = "0.21.0"
 bb8 = "0.8"
 blake3 = "1.3.3"
@@ -55,7 +53,7 @@ literally = "0.1.3"
 maud = { version = "0.24", features = ["actix-web"] }
 mimalloc = { version = "0.1", optional = true }
 mime = "0.3.16"
-moka = {version = "0.10", features = ["future"] }
+moka = { version = "0.10", features = ["future"] }
 nanoid = "0.4.0"
 num_cpus = "1.15.0"
 once_cell = "1.17.1"
@@ -83,6 +81,7 @@ uuid = { version = "1.3.0", features = ["serde", "v4"] }
 # First party crates
 api_models = { version = "0.1.0", path = "../api_models" }
 common_utils = { version = "0.1.0", path = "../common_utils" }
+external_services = { version = "0.1.0", path = "../external_services" }
 masking = { version = "0.1.0", path = "../masking" }
 redis_interface = { version = "0.1.0", path = "../redis_interface" }
 router_derive = { version = "0.1.0", path = "../router_derive" }
@@ -90,7 +89,7 @@ router_env = { version = "0.1.0", path = "../router_env", features = ["log_extra
 storage_models = { version = "0.1.0", path = "../storage_models", features = ["kv_store"] }
 
 [build-dependencies]
-router_env = { version = "0.1.0", path = "../router_env", default-features = false } 
+router_env = { version = "0.1.0", path = "../router_env", default-features = false }
 
 [dev-dependencies]
 actix-http = "3.3.1"
