+++ crates/router/src/services.rs
@@ -1,8 +1,6 @@
 pub mod api;
 pub mod authentication;
 pub mod encryption;
-#[cfg(feature = "kms")]
-pub mod kms;
 pub mod logger;
 
 use std::sync::{atomic, Arc};
